llamacpp: add a vulkan flavor
Add a Containerfile that builds chatbot with llama.cpp enabling the
vulkan backend and install the patched mesa packages to the image. This
is primarily intended to be used on macOS with podman and krunkit, but
can also be used on Linux with a VMM exposing a Venus-capable virtio-gpu
device.

Signed-off-by: Sergio Lopez <slp@redhat.com>