Add tmux to the bootc image
Many commands that are run for SDG and training can take a lot of time,
so there is a risk to have a network disconnection during the task. With
`tmux`, users have the ability to detach the jobs from their SSH session
and let the tasks run for a very long time.

Signed-off-by: Fabien Dupont <fdupont@redhat.com>