Make growfs conditional
growfs is created by Makefile and CI does not use it. Also if I'm not misktaken growfs is only used for disk images creation.
By changing this, growfs file will only be created when Makefile is runningso CI pipelines can build the Containerfile and growfs can be also used when needed.

Signed-off-by: Enrique Belarte Luque <ebelarte@redhat.com>