ilab-wrapper: support execution by non-sudoers
# Background

df8885777d2257a6271ca1a6461673d1d96e99d6

# Issue

Running the script requires the user to be a full sudoer, which is not
a great requirement to have.

# Solution

Create a new group, `ilab`, and add a sudoers rule that allows users in
the `ilab` group to run the wrapper-script as sudo without a password
prompt.

Users will not actually run the script as sudo directly, but the script
will re-run itself as sudo.

Signed-off-by: Omer Tuchfeld <omer@tuchfeld.dev>